Fixing and optimizing tests.

+++ src/landmark_converter.cpp
@@ -34,12 +34,12 @@ std::vector<FaceLandmark> LandmarkConverter::pfldToWflw(const std::vector<FaceLa
             if (pfld_idx < static_cast<int>(pfld_landmarks.size()))
             {
                 // Direct mapping available
-                wflw_landmarks.emplace_back(FaceLandmark{wflw_idx, pfld_landmarks[pfld_idx].p});
+                wflw_landmarks.emplace_back(FaceLandmark{static_cast<unsigned int>(wflw_idx), pfld_landmarks[pfld_idx].p});
             }
             else
             {
                 // Should not happen with valid mapping, but add safeguard
-                wflw_landmarks.emplace_back(FaceLandmark{wflw_idx, math_utils::Point3D(0, 0, 0)});
+                wflw_landmarks.emplace_back(FaceLandmark{static_cast<unsigned int>(wflw_idx), math_utils::Point3D(0, 0, 0)});
             }
         }
         else
@@ -48,12 +48,12 @@ std::vector<FaceLandmark> LandmarkConverter::pfldToWflw(const std::vector<FaceLa
             // For now, use first 98 landmarks as fallback (our current approach)
             if (wflw_idx < static_cast<int>(pfld_landmarks.size()))
             {
-                wflw_landmarks.emplace_back(FaceLandmark{wflw_idx, pfld_landmarks[wflw_idx].p});
+                wflw_landmarks.emplace_back(FaceLandmark{static_cast<unsigned int>(wflw_idx), pfld_landmarks[wflw_idx].p});
             }
             else
             {
                 // Fallback to origin - should not happen
-                wflw_landmarks.emplace_back(FaceLandmark{wflw_idx, math_utils::Point3D(0, 0, 0)});
+                wflw_landmarks.emplace_back(FaceLandmark{static_cast<unsigned int>(wflw_idx), math_utils::Point3D(0, 0, 0)});
             }
         }
     }
@@ -90,7 +90,7 @@ std::vector<FaceLandmark> LandmarkConverter::wflwToPfld(const std::vector<FaceLa
     // Initialize all landmarks to origin
     for (int i = 0; i < 106; ++i)
     {
-        pfld_landmarks[i] = FaceLandmark{i, math_utils::Point3D(0, 0, 0)};
+        pfld_landmarks[i] = FaceLandmark{static_cast<unsigned int>(i), math_utils::Point3D(0, 0, 0)};
     }
 
     // Map existing landmarks
@@ -102,7 +102,7 @@ std::vector<FaceLandmark> LandmarkConverter::wflwToPfld(const std::vector<FaceLa
             int pfld_idx = it->second;
             if (pfld_idx < 106)
             {
-                pfld_landmarks[pfld_idx] = FaceLandmark{pfld_idx, wflw_landmarks[wflw_idx].p};
+                pfld_landmarks[pfld_idx] = FaceLandmark{static_cast<unsigned int>(pfld_idx), wflw_landmarks[wflw_idx].p};
             }
         }
     }
@@ -499,12 +499,12 @@ std::vector<FaceLandmark> LandmarkConverter::interpolateLandmarks(const std::vec
         if (count > 0)
         {
             interpolated.emplace_back(
-                FaceLandmark{missing_idx, math_utils::Point3D(sum_x / count, sum_y / count, sum_z / count)});
+                FaceLandmark{static_cast<unsigned int>(missing_idx), math_utils::Point3D(sum_x / count, sum_y / count, sum_z / count)});
         }
         else
         {
             // Fallback: use center point or zero
-            interpolated.emplace_back(FaceLandmark{missing_idx, math_utils::Point3D(0, 0, 0)});
+            interpolated.emplace_back(FaceLandmark{static_cast<unsigned int>(missing_idx), math_utils::Point3D(0, 0, 0)});
         }
     }
